pyproject.toml
[project]
name = "student-homework-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.7.2",
    "fastapi>=0.116.1",
    "litellm>=1.76.2",
    "pillow>=11.3.0",
    "pydantic-settings>=2.10.1",
    "uvicorn>=0.35.0",
]
